Re-org merge stfn utils
