+++ packages/beacon-state-transition/src/altair/index.ts
@@ -2,12 +2,12 @@ export * from "./block";
 export * from "./epoch";
 export * from "./upgradeState";
 
-// re-export phase0 lodestar types for ergonomic usage downstream
+// re-export altair lodestar types for ergonomic usage downstream
 // eg:
 //
 // import {altair} from "@chainsafe/lodestar-beacon-state-transition";
 //
-// altair.processDeposit(...)
+// altair.processSyncAggregate(...)
 //
 // const x: altair.BeaconState;
 export * from "@chainsafe/lodestar-types/altair";
