+++ packages/types/merge.d.ts
@@ -1,0 +1,1 @@
+export * from "./lib/merge/types";
